terzo commit windows

+++ src/main/java/it/bitify/libreria/repository/CategoriaRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepo extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepo extends JpaRepository<Categoria, Long> {
 }
