+++ src/main/java/it/bitify/libreria/repository/CorsoRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Corso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CorsoRepo extends JpaRepository<Corso, Integer> {
+public interface CorsoRepo extends JpaRepository<Corso, Long> {
 }
