+++ src/main/java/it/bitify/libreria/repository/StudenteRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Studente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudenteRepo extends JpaRepository<Studente, Integer> {
+public interface StudenteRepo extends JpaRepository<Studente, Long> {
 }
