+++ src/main/java/it/bitify/libreria/repository/TesseraRepo.java
@@ -3,5 +3,5 @@ package it.bitify.libreria.repository;
 import it.bitify.libreria.entity.Tessera;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TesseraRepo extends JpaRepository<Tessera, Integer> {
+public interface TesseraRepo extends JpaRepository<Tessera, Long> {
 }
